pyproject.toml
[tool.poetry]
name = "backend_dacot"
version = "0.1.0"
description = ""
authors = ["dacot@uoct.cl"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
mongoengine = "^0.23.0"
fastapi-mail = "^0.3.7"
python-magic = "^0.4.22"
requests = "^2.25.1"
mongomock = "^3.22.1"
oauth2client = "^4.1.3"
google-auth = "^1.30.0"
pymongo = "^3.11.3"
graphene-mongo = "^0.2.13"
fastapi-jwt-auth = "^0.5.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
aiohttp = "^3.7.4"
gql = "^2.0.0"
gql-query-builder = "^0.1.5"
loguru = "^0.5.3"
sendgrid = "^6.7.0"
Authlib = "^0.15.4"
itsdangerous = "^2.0.1"
pytest = "^6.2.4"
coverage = "^5.5"
dacot-models = {path = "dacot_models"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
